+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Additions
 - Added InputHandler.InvertPressBehavior
 - Added ReverseInput, ReverseOutput and AndThen to Easings
 - Added an Enum constructor to GenericInput
+- Added RandomPitchModifier and GetRandomPitch to SoundEffectInfo
 
 Improvements
 - Allow comparing Keybind and Combination based on the amount of modifiers they have